Added password field to create site, as you need to enter your password at this stage. The switch is now more of a show/hide field.
Updated command.rs with just some print statements to remind me to implement github and password integration, but not suported as of yet.

Added password field to NewSite struct and some traits in lib.rs.

Updated create_site to send the password instead of a true/false, and print out the response, and removed verbose logging in the read_object_response, response now responds with the full text value of the response instead of the status code, as the status is much less helpful.