Added warning class to tailwind
Updated create site to include an alert dialog for the password section, un-commented out RSS and then applied a hidden class to everything so that things that aren't enabled/tested yet aren't visible. Also added a file upload button for the favicon.

Added a file picker util file

Added some documentation comments

Added an openWiki function in utils to quickly open the driftwood wiki for internal app links

Added shell and dialog to the tauri allow list

Added check_site_dir and move_favicon_to_site_dir fn in lib.rs

Updated create_site fn in Netlify.rs to handle sending a password or not depending on if password is enabled, creating a site dir, and moving the favicon to the site dir if a favicon file is supplied.

Updated the read_object_response error handling so we get the actual error message from netlify back.