Added handlePostReturnClick which returns from the post create screen to the edit site screen. Passed this to the Post component.
Added favicon_file to the schema and structures and type for the edit site screen.

Changed "requried" fields from booleans to strings in typescript. Technically they are string arrays but, doesn't really matter, we never pass values to this for the client, it's a field only used for  deploying sites. tbf, should have a "deploy site" struct and a "site details" struct.

Currently in get_site_details in commands updated to return a custom json response of the site details that includes the favicon file. This file just needs to be retrieved from the site's file dir. Will just to a 'grep' for .ico files in the site dir to find it. Then need to add the file picker to edit site screen.