Updated create site's site name schema, and edit site, so the site name can only be letters, numbers, and dashes.
Commented out all password fields and mentions in edit_site.

Added site id to edit site, so you can see what you're working on.

Added filepicked to favicon for edit site, and the favicon chosen will be placed in the site dir.

Updated file hashing so we send the favicon to get hashed as well.

Updated the delete site alert, so it just informs users this is unsupported currently.

refresh sites on creation and updating.

in get site details, also retrieve the favicon file if it exists. Added favicon_file to SiteDetails for ease of transportation.

Made site dirs just the id, not the name, in case the name is changed. This also means removing the .name from anywhere we build the site path.